Restructure and rename classes/namespaces for consistency
