+++ Console.CodingTracker.tonyissa/UserInterface.cs
@@ -1,5 +1,6 @@
 using CodingTracker.Dates;
 using CodingTracker.Database;
+using CodingTracker.Input;
 
 namespace CodingTracker.UserInterface
 {
@@ -12,22 +13,31 @@ namespace CodingTracker.UserInterface
             {
                 Console.Clear();
                 Console.WriteLine("Welcome to my coding tracker!");
-                Console.WriteLine("Press C to create a record, U to update, D to delete, V to view all, and 0 to exit.");
+                Console.WriteLine("Press C to create a record, U to update, D to delete, V to view all, R to record, and 0 to exit.");
 
                 var input = Console.ReadLine() ?? "0";
                 input = input.Trim().ToLower();
 
-                if (input == "0") return;
-                else if (input == "c") InitCreateMenu();
-                else if (input == "d") break;
-                else if (input == "u") break;
-                else if (input == "v") GetAllEntries();
-                else
+                switch (input)
                 {
-                    Console.WriteLine("Input not recognized. Please try again.");
-                    continue;
+                    case "0":
+                        break;
+                    case "c":
+                        InitCreateMenu();
+                        break;
+                    case "d":
+                        break;
+                    case "u":
+                        break;
+                    case "v":
+                        PrintAllEntries();
+                        break;
+                    case "r":
+                        break;
+                    default:
+                        Console.WriteLine("Input not recognized. Please try again.");
+                        continue;
                 }
-
             }
         }
 
@@ -56,7 +66,7 @@ namespace CodingTracker.UserInterface
             }
         }
 
-        public static void GetAllEntries()
+        public static void PrintAllEntries()
         {
             var results = DatabaseController.GetList();
             var index = 1;
@@ -76,19 +86,19 @@ namespace CodingTracker.UserInterface
 
         public static (DateTime, DateTime)? GetAllDateTimes()
         {
-            var startDate = GetDateInput(true);
+            var startDate = InputHelper.GetDateInput(true);
             if (startDate == "0") return null;
             else if (startDate == "") startDate = DateHelper.GetCurrentDate();
 
-            var startTime = GetTimeInput(true);
+            var startTime = InputHelper.GetTimeInput(true);
             if (startTime == "0") return null;
             else if (startTime == "") startTime = DateHelper.GetCurrentTime();
 
-            var endDate = GetDateInput(false);
+            var endDate = InputHelper.GetDateInput(false);
             if (endDate == "0") return null;
             else if (endDate == "") endDate = DateHelper.GetCurrentDate();
 
-            var endTime = GetTimeInput(false);
+            var endTime = InputHelper.GetTimeInput(false);
             if (endTime == "0") return null;
             else if (endTime == "") endTime = DateHelper.GetCurrentTime();
 
@@ -97,48 +107,10 @@ namespace CodingTracker.UserInterface
 
             if (!DateHelper.CompareDates(date1, date2))
             {
-                throw new ArgumentException("Date 1 equal or later than date 2. Press any key to continue and try again.");
+                throw new ArgumentException("Date 1 equal or later than date 2.");
             }
 
             return (date1, date2);
-        }
-
-        public static string GetDateInput(bool start)
-        {
-            string s = start ? "start" : "end";
-            Console.WriteLine($"Please input the {s} date in the M/D/YYYY format, type 0 to quit, or press enter for the current date.");
-
-            while (true)
-            {
-                var input = Console.ReadLine() ?? "0";
-                input = input.Trim().ToLower();
-
-                if (input == "0" || input == string.Empty || DateHelper.ValidateDateFormat(input))
-                {
-                    return input;
-                }
-
-                Console.WriteLine("Invalid date.");
-            }
-        }
-
-        public static string GetTimeInput(bool start)
-        {
-            string s = start ? "start" : "end";
-            Console.WriteLine($"Please input the {s} time in either hh:mm tt (12-hour time) format, HH:MM (24-hour time) format, type 0 to quit, or press enter for the current time.");
-
-            while (true)
-            {
-                var input = Console.ReadLine() ?? "0";
-                input = input.Trim().ToLower();
-
-                if (input == "0" || input == string.Empty || DateHelper.ValidateTimeFormat(input))
-                {
-                    return input;
-                }
-
-                Console.WriteLine("Invalid time.");
-            }
         }
     }
 }
